Refatorando serviço de token

+++ API-C#/WorkFlow.Service/Interface/IAccountService.cs
@@ -6,7 +6,7 @@ namespace WorkFlow.Service.Interface
 {
     public interface IAccountService
     {
-        Task<ResponseDTO> CreateOrUpdateUser(long? id, UserDTO user);
+        Task<ResponseDTO> CreateOrUpdateUser(int? id, UserDTO user);
         Task<ResponseDTO> AddUserInRole(User user, RoleName role);
         Task<ResponseDTO> GetUserByUserNameAsync(string userName);
         Task<ResponseDTO> Login(UserLoginDTO userDTO);
